skia_bench: fix crash due to memory issue
currently the BenchTimer.cpp is pulling in the class
definintions for its underlying timers from BenchSysTimer_c.[h/c].

However, the Android.mk links it with BenchSysTimer_posix.c. This
is causing a crash when the destructor calls delete on the object.

This patch ensures that we pull in the header required by Android.mk.

NOTE: simply defining SK_BUILD_FOR_UNIX does not work because that
causes other parts of skia_bench to using things link GL that we don't
have access to in an android build.

Change-Id: If3d643d07de0ea9ae0f34ff7a009527ea2c323a5
Author: Andy Doan <andy.doan@linaro.org>
Signed-off-by: Andy Doan <andy.doan@linaro.org>
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>

Conflicts:
	bench/BenchTimer.cpp